% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGB_prep.R
\name{AGB_prep}
\alias{AGB_prep}
\title{Prepare Landtrendr AGB data for the SDA workflow.}
\usage{
AGB_prep(
  Site_Info,
  Start_Date,
  End_Date,
  AGB_dir = "/projectnb/dietzelab/dongchen/Multi-site/download_500_sites/AGB",
  OutDir = NULL,
  Export_CSV = TRUE,
  Allow_download = FALSE
)
}
\arguments{
\item{Site_Info}{Bety list of site info including site_id, lon, and lat.}

\item{Start_Date}{Start date of SDA workflow.}

\item{End_Date}{End date of SDA workflow.}

\item{AGB_dir}{Where the Landtrendr AGB data can be accessed.}

\item{OutDir}{Where the final CSV file will be stored.}

\item{Export_CSV}{Decide if we want to export the CSV file.}

\item{Allow_download}{If data is missing, should we download the missing data?}
}
\value{
A data frame containing AGB median and sd for each site and each time step.
}
\description{
Prepare Landtrendr AGB data for the SDA workflow.
}
\author{
Dongchen Zhang
}
