% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_remote_process.R
\name{call_remote_process}
\alias{call_remote_process}
\title{call_remote_process}
\usage{
call_remote_process(settings)
}
\arguments{
\item{settings}{PEcAn settings list containing remotedata tags: siteid, sitename, raw_mimetype, raw_formatname, geofile, outdir, start, end, source, collection, scale, projection, qc, algorithm, credfile, pro_mimetype, pro_formatname, out_get_data, out_process_data}
}
\description{
call the Python - remote_process from PEcAn and store the output in BETY
}
\examples{
\dontrun{
call_remote_process(settings)
}
}
\author{
Ayush Prasad
}
