% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_process.R
\name{collapse_remote_data}
\alias{collapse_remote_data}
\title{collapse_remote_data}
\usage{
collapse_remote_data(output, out_process_data, latlon)
}
\arguments{
\item{output}{output list from rp_control}

\item{out_process_data}{type of processed data}

\item{latlon}{latlon}
}
\description{
collapses remote data (currently LAI only)
}
\examples{
\dontrun{
collapse_remote_data(
  output,
  out_process_data,
  latlon)
}
}
\author{
Istem Fer
}
