% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_process.R
\name{construct_raw_filename}
\alias{construct_raw_filename}
\title{construct_raw_filename}
\usage{
construct_raw_filename(
  collection,
  siteid,
  scale = NULL,
  projection = NULL,
  qc = NULL
)
}
\arguments{
\item{collection}{collection or product requested from the source}

\item{scale}{scale, NULL by default}

\item{projection}{projection, NULL by default}

\item{qc}{qc_parameter, NULL by default}

\item{sitename}{site name}

\item{source}{source of the remote data}
}
\value{
raw_file_name
}
\description{
construct the raw file name
}
\examples{
\dontrun{
raw_file_name <- construct_raw_filename(
  sitename="Reykjavik",
  source="gee",
  collection="s2",
  scale=10.0
  projection=NULL
  qc=1.0)
}
}
\author{
Ayush Prasad
}
