% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_process.R
\name{construct_remotedata_filename}
\alias{construct_remotedata_filename}
\title{construct_remotedata_filename}
\usage{
construct_remotedata_filename(
  source,
  collection,
  siteid,
  scale = NULL,
  projection = NULL,
  qc = NULL,
  algorithm = NULL,
  out_process_data = NULL
)
}
\arguments{
\item{source}{source}

\item{collection}{collection or product requested from the source}

\item{siteid}{shortform of siteid}

\item{scale}{scale, NULL by default}

\item{projection}{projection, NULL by default}

\item{qc}{qc_parameter, NULL by default}

\item{algorithm}{algorithm name to process data, NULL by default}

\item{out_process_data}{variable name requested for the processed file, NULL by default}
}
\value{
remotedata_file_names
}
\description{
construct remotedata module file names
}
\examples{
\dontrun{
remotedata_file_names <- construct_remotedata_filename(
  source="gee",
  collection="s2",
  siteid="0-721",
  scale=10.0
  projection=NULL
  qc=1.0,
  algorithm="snap",
  out_process_data="lai")
}
}
\author{
Ayush Prasad
}
