% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.AGB.R
\name{download.AGB}
\alias{download.AGB}
\title{download.AGB}
\usage{
download.AGB(outdir, product_dates = NULL,
  temporal_resolution = "annual", product_source = "OSU",
  prodcut_version = NULL, con = NULL, run_parallel = TRUE,
  ncores = NULL, overwrite = FALSE)
}
\arguments{
\item{outdir}{Where to place output}

\item{product_dates}{What dates to download and extract}

\item{temporal_resolution}{Optional. What temporal resolution to download. Current options: "annual"}

\item{product_source}{The name of the desired AGB prodcut.  Current options: "OSU"}

\item{prodcut_version}{Optional. For some products (e.g. OSU) there are multiple versions.}

\item{con}{Optional database connection. If specified then the code will check to see}

\item{run_parallel}{Logical. Download and extract files in parallel?}

\item{ncores}{Optional. If run_parallel=TRUE how many cores to use?  If left as NULL will select max number -1}

\item{overwrite}{Logical. Overwrite existing files and replace with new versions}

\item{generate_plots}{Optional. Create output diagnostic plots in outdir? --- ACTUALLY PUT THIS IN extract_ABG}
}
\description{
download.AGB
}
\examples{
\dontrun{
outdir <- "~/scratch"
product_dates <- c(1990,1991)
product_source = "OSU"
prodcut_version = "v1"
PEcAn.data.remote::download.AGB(outdir=outdir, product_dates = product_dates, 
prodcut_version = prodcut_version)
}

}
\author{
Shawn Serbin
}
