% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandTrendr.AGB.R
\name{download.LandTrendr.AGB}
\alias{download.LandTrendr.AGB}
\title{download.LandTrendr.AGB}
\usage{
download.LandTrendr.AGB(
  outdir,
  target_dataset = "biomass",
  product_dates = NULL,
  product_version = "v1",
  con = NULL,
  run_parallel = TRUE,
  ncores = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{outdir}{Where to store the output - the downloaded LandTrendr AGB raster data}

\item{target_dataset}{Use these argument to select which LandTrendr dataset to download. 
For v1 the default is "biomass" For v2 you should use "biomassfiaald"}

\item{product_dates}{What data product dates to download}

\item{product_version}{Optional. LandTrend AGB is provided with three versions, 
v0, v1, and v2 (latest version)}

\item{con}{Optional database connection. If specified then the code will check to see 
if the file already exists in PEcAn before downloading, and will also create a database 
entry for new downloads - NOT YET IMPLEMENTED}

\item{run_parallel}{Logical. Download and extract files in parallel?}

\item{ncores}{Optional. If run_parallel=TRUE how many cores to use?  If left as NULL will select max number -1}

\item{overwrite}{Logical. Overwrite existing files and replace with new versions}
}
\value{
data.frame summarizing the results of the function call
}
\description{
download.LandTrendr.AGB
}
\examples{
\dontrun{
outdir <- "~/scratch/abg_data/"
product_dates <- c(1990, 1991, 1995)  # using discontinous, or specific years
product_dates2 <- seq(1992, 1995, 1)  # using a date sequence for selection of years
product_version = "v2"
target_dataset = "biomassfiaald"

results <- PEcAn.data.remote::download.LandTrendr.AGB(outdir=outdir,
           target_dataset=target_dataset,
           product_dates = product_dates, 
           product_version = product_version)

results <- PEcAn.data.remote::download.LandTrendr.AGB(outdir=outdir, 
           target_dataset=target_dataset,
           product_dates = product_dates2, 
           product_version = product_version)
}

}
\author{
Shawn Serbin
}
