% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.thredds.R
\name{download.thredds}
\alias{download.thredds}
\alias{download_thredds}
\title{download.thredds}
\usage{
download_thredds(site_info, dates, varid, dir_url, data_url,
  run_parallel = FALSE, outdir = NULL)
}
\arguments{
\item{site_info}{list of information with the site_id, site_info, lat, lon, and time_zone. Derived from BETY using a PEcAn .xml settings file with site information. Can use the get_site_info function to generate this list.}

\item{dates}{vector of start and end date for dataset as YYYYmmdd, YYYY-mm-dd, YYYYjjj, or date object.}

\item{varid}{character vector of shorthand variable name. i.e. LAI}

\item{dir_url}{catalog url of data from ncei.noaa.gov/thredds website}

\item{data_url}{opendap url of data from ncei.noaa.gov/thredds website}

\item{run_parallel}{Logical. Download and extract files in parallel?}

\item{outdir}{file location to place output}
}
\value{
data.frame summarize the results of the function call
}
\description{
download.thredds
}
\author{
Bailey Morrison
}
