% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.thredds.R
\name{extract_thredds_nc}
\alias{extract_thredds_nc}
\title{extract_thredds_nc}
\usage{
extract_thredds_nc(site_info, url, varid)
}
\arguments{
\item{site_info}{list containing site_id, site_name, lat, lon, time_zone.
Derived from BETY using a PEcAn .xml settings file with site information.
Can use the get_site_info function to generate this list.}

\item{url}{a THREDDS url of a .nc file to extract data from.}

\item{varid}{character vector of shorthand variable name. i.e. LAI}
}
\value{
a dataframe with the values for each date/site combination
 from a THREDDS file
}
\description{
extract_thredds_nc
}
\examples{
\dontrun{
settings <- PEcAn.settings::read.settings("/path/to/pecan.xml")
site_info <- get_site_info(settings)
thredds_url = paste0( # breaking up long URL for readability
  "https://www.ncei.noaa.gov/thredds/dodsC/cdr/lai/files/1995/",
  "AVHRR-Land_v005_AVH15C1_NOAA-14_19950201_c20180831220722.nc")
output <- extract_thredds_nc(
  site_info = site_info,
  url = thredds_url,
  varid = "LAI")
}
}
\author{
Bailey Morrison
}
