% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_image_tiles.R
\name{gdal_translate}
\alias{gdal_translate}
\title{gdal_translate}
\usage{
gdal_translate(from, to)
}
\arguments{
\item{from}{character: subdataset name. 
Generated from the `get_subdatasets` function.}

\item{to}{character: physical path to the output file.}
}
\description{
This function provides tool the gdal_translate execution.
}
\author{
Dongchen Zhang
}
