% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{GEDI_AGB_download}
\alias{GEDI_AGB_download}
\title{Download GEDI AGB data for the GEDI AGB extract function.}
\usage{
GEDI_AGB_download(
  start_date,
  end_date,
  outdir,
  extends,
  nfile.min = 0,
  nrow.min = 0,
  gradient = 0.001
)
}
\arguments{
\item{start_date}{start date for filtering out the existing CSV file.}

\item{end_date}{end date for filtering out the existing CSV file.}

\item{outdir}{Where the final CSV file will be stored.}

\item{extends}{the XY box (in degrees) for downloading GEDI AGB file.}

\item{nfile.min}{the minimum required file number to be downloaded and extracted.}

\item{nrow.min}{the minimum required observation number to be extracted.}

\item{gradient}{the gradient for iteratively enlarge the extends if the nfile.min or nrow.min are not reached.}
}
\value{
A data frame containing AGB and sd for the target spatial and temporal extends.
}
\description{
Download GEDI AGB data for the GEDI AGB extract function.
}
\author{
Dongchen Zhang
}
