% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{GEDI_AGB_extract}
\alias{GEDI_AGB_extract}
\title{Extract GEDI AGB data for the GEDI AGB prep function.}
\usage{
GEDI_AGB_extract(
  site_info,
  start_date,
  end_date,
  outdir,
  nfile.min = 0,
  nrow.min = 0,
  buffer = 0.01
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{start_date}{start date for filtering out the existing CSV file.}

\item{end_date}{end date for filtering out the existing CSV file.}

\item{outdir}{Where the final CSV file will be stored.}

\item{nfile.min}{the minimum required file number to be downloaded and extracted.}

\item{nrow.min}{the minimum required observation number to be extracted.}

\item{buffer}{buffer distance for locate GEDI AGB searching box (default is 0.01 ~ 1 km)}
}
\value{
A list of AGB data frames for each site.
}
\description{
Extract GEDI AGB data for the GEDI AGB prep function.
}
\author{
Dongchen Zhang
}
