% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{GEDI_AGB_plot}
\alias{GEDI_AGB_plot}
\title{Plot GEDI AGB observations around the target site.}
\usage{
GEDI_AGB_plot(outdir, site.id, start_date, end_date)
}
\arguments{
\item{outdir}{Where the plot PNG file will be stored.}

\item{site.id}{Unique ID for the target site.}

\item{start_date}{start date (date with YYYY-MM-DD format) for filtering out the existing CSV file.}

\item{end_date}{end date (date with YYYY-MM-DD format) for filtering out the existing CSV file.}
}
\description{
Plot GEDI AGB observations around the target site.
}
\author{
Dongchen Zhang
}
