% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{GEDI_L4A_2_mean_var}
\alias{GEDI_L4A_2_mean_var}
\title{Aggregate AGB mean and uncertainty from the GEDI level4A tiles.}
\usage{
GEDI_L4A_2_mean_var(
  site_info,
  which.point.in.which.file,
  buffer = 0.005,
  cores = parallel::detectCores()
)
}
\arguments{
\item{site_info}{List: list of site info including site_id, site_name, lon, and lat.}

\item{which.point.in.which.file}{List: lists containing physical paths of GEDI tiles that intercept the each site.}

\item{buffer}{Numeric: buffer distance (in degree) that is used to create the bounding box (default is 0.005 [~ 500 m]).}

\item{cores}{Numeric: numbers of core to be used for the parallel computation. The default is the maximum current CPU number.}
}
\value{
A list containing AGB mean and standard deviation for each site.
}
\description{
Aggregate AGB mean and uncertainty from the GEDI level4A tiles.
}
