% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{GEDI_L4A_Finder}
\alias{GEDI_L4A_Finder}
\title{Detect which GEDI level 4A tiles intercept which site.}
\usage{
GEDI_L4A_Finder(
  files,
  site_info,
  buffer = 0.005,
  cores = parallel::detectCores()
)
}
\arguments{
\item{files}{Character: full paths of GEDI level 4A tiles.}

\item{site_info}{List: list of site info including site_id, site_name, lon, and lat.}

\item{buffer}{Numeric: buffer distance (in degree) that is used to create the bounding box (default is 0.005 [~ 500 m]).}

\item{cores}{Numeric: numbers of core to be used for the parallel computation. The default is the maximum current CPU number.}
}
\value{
A list containing physical paths of GEDI tiles that intercept the each site.
}
\description{
Detect which GEDI level 4A tiles intercept which site.
}
