% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{GEDI_L4A_Finder_batch}
\alias{GEDI_L4A_Finder_batch}
\title{Submit jobs through `qsub` for the `GEDI_L4A_Finder` function.}
\usage{
GEDI_L4A_Finder_batch(
  files,
  outdir,
  site_info,
  num.folder,
  buffer = 0.005,
  cores = parallel::detectCores(),
  prerun = NULL
)
}
\arguments{
\item{files}{Character: full paths of GEDI level 4A tiles.}

\item{outdir}{Character: the physical path within which the batch job folders will be created.}

\item{site_info}{List: list of site info including site_id, site_name, lon, and lat.}

\item{num.folder}{Numeric: the number of batch folders to be created.}

\item{buffer}{Numeric: buffer distance (in degree) that is used to create the bounding box (default is 0.005 [~ 500 m]).}

\item{cores}{Numeric: numbers of core to be used for the parallel computation. The default is the maximum current CPU number.}

\item{prerun}{Character: a vector of strings that will be executed beforehand. The default is NULL.}
}
\value{
A list containing physical paths of GEDI tiles that intercept the each site.
}
\description{
Submit jobs through `qsub` for the `GEDI_L4A_Finder` function.
}
