% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.thredds.R
\name{get_site_info}
\alias{get_site_info}
\title{get_site_info}
\usage{
get_site_info(settings)
}
\arguments{
\item{settings}{a PEcAn settings object}
}
\value{
a list of site information derived from BETY using a pecan .xml
 settings file with site_id, site_name, lat, lon, and time_zone.
}
\description{
get_site_info
}
\examples{
\dontrun{
settings <- PEcAn.settings::read.settings("/path/to/pecan.xml")
site_info <- get_site_info(settings)
}
}
\author{
Bailey Morrison
}
