% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_image_tiles.R
\name{get_subdatasets}
\alias{get_subdatasets}
\title{get_subdatasets.}
\usage{
get_subdatasets(in_path)
}
\arguments{
\item{in_path}{character: physical path to the image file.}
}
\description{
This function provides tool for reading band names of remote sensing image.
}
\author{
Dongchen Zhang
}
