% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASA_DAAC_download.R
\name{getnetrc}
\alias{getnetrc}
\title{Set NASA DAAC credentials to the .netrc file.}
\usage{
getnetrc(dl_path)
}
\arguments{
\item{dl_path}{Character: physical path to the .netrc credential file.}
}
\description{
Set NASA DAAC credentials to the .netrc file.
}
\author{
Dongchen Zhang
}
