% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{grid2netcdf}
\alias{grid2netcdf}
\title{grid2netcdf}
\usage{
grid2netcdf(gdata, date = "9999-09-09", outfile = "out.nc")
}
\arguments{
\item{gdata}{gridded data to write out}

\item{date}{currently ignored; date(s) from \code{gdata} are used instead}

\item{outfile}{name for generated netCDF file.}
}
\value{
writes netCDF file
}
\description{
Write gridded data to netcdf file
}
\author{
David LeBauer
}
