% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAI_prep.R
\name{LAI_prep}
\alias{LAI_prep}
\title{Prepare MODIS LAI data for the SDA workflow.}
\usage{
LAI_prep(
  Site_Info,
  Start_Date,
  End_Date,
  Time_Step = list(unit = "year", num = 1),
  NCore = NULL,
  OutDir = NULL,
  Search_Window = 30,
  Export_CSV = FALSE
)
}
\arguments{
\item{Site_Info}{Bety list of site info including site_id, lon, and lat.}

\item{Start_Date}{Start date of SDA workflow.}

\item{End_Date}{End date of SDA workflow.}

\item{Time_Step}{A list containing time step and number of time step,
which allows time step to be any years or days.}

\item{NCore}{Number of CPU to be used for LAI extraction.}

\item{OutDir}{Where the final CSV file will be stored.}

\item{Search_Window}{search window for locate available LAI values.}

\item{Export_CSV}{Decide if we want to export the CSV file.}
}
\value{
A data frame containing LAI and sd for each site and each time step.
}
\description{
Prepare MODIS LAI data for the SDA workflow.
}
\author{
Dongchen Zhang
}
