% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landtrendr_AGB_prep.R
\name{Landtrendr_AGB_prep}
\alias{Landtrendr_AGB_prep}
\title{Prepare Landtrendr AGB data for the SDA workflow.}
\usage{
Landtrendr_AGB_prep(
  site_info,
  start_date,
  end_date,
  time_points,
  AGB_indir,
  outdir = NULL,
  export_csv = TRUE,
  allow_download = FALSE,
  buffer = NULL,
  skip_buffer = TRUE
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{start_date}{Start date of SDA workflow.}

\item{end_date}{End date of SDA workflow.}

\item{time_points}{A vector contains each time point within the start and end date.}

\item{AGB_indir}{Where the Landtrendr AGB data can be accessed.}

\item{outdir}{Where the final CSV file will be stored.}

\item{export_csv}{Decide if we want to export the CSV file.}

\item{allow_download}{If data is missing, should we download the missing data?}

\item{buffer}{buffer area to calculate the min var of AGB data.}

\item{skip_buffer}{flag to skip calculating min var based on buffer area.}
}
\value{
A data frame containing AGB median and sd for each site and each time step.
}
\description{
Prepare Landtrendr AGB data for the SDA workflow.
}
\author{
Dongchen Zhang
}
