% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_image_tiles.R
\name{merge_image_tiles}
\alias{merge_image_tiles}
\title{Merge image tiles to a single image (currently support hdf and tif image format).}
\usage{
merge_image_tiles(
  folder.path,
  keep.files = FALSE,
  image.settings = list(crs = "EPSG:4326", dimension = NULL, ext = NULL, fun = NULL),
  computation = list(GDAL_CACHEMAX = 1000, wm = "80\%", NUM_THREADS = 16, COMPRESS =
    "DEFLATE")
)
}
\arguments{
\item{folder.path}{character: physical path to the folder that contains all the image tiles.}

\item{keep.files}{Boolean: if we want to keep the image tiles at the end.}

\item{image.settings}{list: settings used during exporting merged image.
Such as image coordinate system (crs), dimension, extents (ext), and average function (fun).}

\item{computation}{list: settings used for configuring computation.
Such as maximum memory per CPU (GDAL_CACHEMAX), percentage of total memory (wm),
number of CPUs (NUM_THREADS), compress method (COMPRESS).}
}
\value{
character: file path to the merged GeoTIFF file.
}
\description{
Merge image tiles to a single image (currently support hdf and tif image format).
}
\details{
Please refer to the gdalwarp manual for more details
https://gdal.org/en/stable/programs/gdalwarp.html
}
\author{
Dongchen Zhang
}
