% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS_LAI_prep.R
\name{MODIS_LAI_prep}
\alias{MODIS_LAI_prep}
\title{Prepare MODIS LAI data for the SDA workflow.}
\usage{
MODIS_LAI_prep(
  site_info,
  time_points,
  outdir = NULL,
  search_window = 30,
  export_csv = FALSE,
  sd_threshold = 20,
  skip_download = FALSE,
  boundary = NULL
)
}
\arguments{
\item{site_info}{list: Bety list of site info including site_id, lon, and lat.}

\item{time_points}{character: a vector contains each time point within the start and end date.}

\item{outdir}{character: where the final CSV file will be stored.}

\item{search_window}{numeric: search window for locate available LAI values.}

\item{export_csv}{boolean: decide if we want to export the CSV file.}

\item{sd_threshold}{numeric or character: for filtering out any estimations with unrealistic high standard error, default is 20. The QC check will be skipped if it's set as NULL.}

\item{skip_download}{boolean: determine if we want to use existing LAI.csv file and skip the MODIS LAI download part.}

\item{boundary}{numeric vector or list: the upper and lower quantiles for filtering out noisy LAI values (e.g., c(0.05, 0.95) or list(0.05, 0.95)). The default is NULL.}
}
\value{
A data frame containing LAI and sd for each site and each time step.
}
\description{
Prepare MODIS LAI data for the SDA workflow.
}
\author{
Dongchen Zhang
}
