% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS_LAI_prep.R
\name{MODIS_LAI_ts_filter}
\alias{MODIS_LAI_ts_filter}
\title{Prepare MODIS LAI data from the NASA DAAC server for the SDA workflow.}
\usage{
MODIS_LAI_ts_filter(lai.csv, boundary = c(0.05, 0.95))
}
\arguments{
\item{lai.csv}{data frame: A data frame containing date, site id, lat, lon, lai, sd, and qc columns.}

\item{boundary}{numeric: A vector contains the upper and lower quantiles for filtering lai records. The default is c(0.05, 0.95).}
}
\value{
A data frame containing date, site id, lat, lon, lai, sd, and qc columns.
}
\description{
Prepare MODIS LAI data from the NASA DAAC server for the SDA workflow.
}
\author{
Dongchen Zhang
}
