% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS_LC_prep.R
\name{MODIS_LC_prep}
\alias{MODIS_LC_prep}
\title{Prepare MODIS land cover data for the SDA workflow.}
\usage{
MODIS_LC_prep(
  site_info,
  time_points,
  outdir = NULL,
  qc.filter = c("000", "001")
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{time_points}{A vector contains each time point within the start and end date.}

\item{outdir}{Where the final CSV file will be stored.}

\item{qc.filter}{values that will pass the QC check. the check will be skipped if it's NULL.}
}
\value{
A data frame containing MODIS land cover types for each site and each time step.
}
\description{
Prepare MODIS land cover data for the SDA workflow.
}
\details{
This function enables the feature of grabbing pre-extracted MODIS LC CSV files such that any site that
has records will be skipped (See Line 33). In more detail, we will be loading the previous `LC.csv` file, which
contains previous extracted land cover records and trying to match that with current requests (location, time).
Any requests that fail the match will be regarded as new extractions and combine with the previous `LC.csv` file.
}
\author{
Dongchen Zhang
}
