% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASA_DAAC_download.R
\name{NASA_CMR_finder}
\alias{NASA_CMR_finder}
\title{Create URL that can be used to request data from NASA DAAC server.}
\usage{
NASA_CMR_finder(doi)
}
\arguments{
\item{doi}{Character: data DOI on the NASA DAAC server, it can be obtained
directly from the NASA ORNL DAAC data portal (e.g., GEDI L4A through
https://daac.ornl.gov/cgi-bin/dsviewer.pl?ds_id=2056).}
}
\value{
A list with each containing corresponding provider and concept ids
given the data doi.
}
\description{
Create URL that can be used to request data from NASA DAAC server.
}
\examples{
\dontrun{
provider_conceptID <- NASA_CMR_finder("10.3334/ORNLDAAC/2183")
}
}
\author{
Dongchen Zhang
}
