% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS_LAI_prep.R
\name{Prep.MODIS.CSV.from.DAAC}
\alias{Prep.MODIS.CSV.from.DAAC}
\title{Prepare MODIS LAI data from the NASA DAAC server for the SDA workflow.}
\usage{
Prep.MODIS.CSV.from.DAAC(
  site_info,
  extent,
  from,
  to,
  download.outdir,
  csv.outdir
)
}
\arguments{
\item{site_info}{list: Bety list of site info including site_id, lon, and lat.}

\item{extent}{numeric: A vector contains the bounding box that covers all sites (West longitude, East longitude, South latitude ,North latitude).}

\item{from}{character: the start time for searching the MODIS products.}

\item{to}{character: the end time for searching the MODIS products.}

\item{download.outdir}{character: Where the MODIS tiles will be stored.}

\item{csv.outdir}{character: Where the final CSV file will be stored.}
}
\value{
A data frame containing LAI and sd for each site and each time step.
}
\description{
Prepare MODIS LAI data from the NASA DAAC server for the SDA workflow.
}
\author{
Dongchen Zhang
}
