% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_AGB_IC_from_2010_global.R
\name{Prep_AGB_IC_from_2010_global}
\alias{Prep_AGB_IC_from_2010_global}
\title{Extract ensemble above ground biomass density from pre-existing GeoTIFF files for the SDA workflow.
Note that, this function only works for those products who have both mean and uncertainty GeoTIFF images prepared.
And it works under the 2010 Global AGB products: DOI: https://doi.org/10.3334/ORNLDAAC/1763.}
\usage{
Prep_AGB_IC_from_2010_global(site_info, paths.list, ens)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{paths.list}{list containing file paths for `mean` and `uncertainty` datasets.}

\item{ens}{ensemble number.}
}
\value{
A data frame containing sampled above ground biomass densities, each column represent each site.
}
\description{
Extract ensemble above ground biomass density from pre-existing GeoTIFF files for the SDA workflow.
Note that, this function only works for those products who have both mean and uncertainty GeoTIFF images prepared.
And it works under the 2010 Global AGB products: DOI: https://doi.org/10.3334/ORNLDAAC/1763.
}
\author{
Dongchen Zhang
}
