% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_AGB_IC_from_geotiff.R
\name{Prep_AGB_IC_from_geotiff}
\alias{Prep_AGB_IC_from_geotiff}
\title{Extract above ensemble ground biomass density from pre-existing GeoTIFF file for the SDA workflow.}
\usage{
Prep_AGB_IC_from_geotiff(site_info, paths.list, ens)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{paths.list}{list containing file paths for `mean` and `uncertainty` datasets.}

\item{ens}{ensemble number.}
}
\value{
A data frame containing sampled above ground biomass densities, each column represent each site.
}
\description{
Extract above ensemble ground biomass density from pre-existing GeoTIFF file for the SDA workflow.
}
\author{
Dongchen Zhang
}
