% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_process.R
\name{remote_process}
\alias{remote_process}
\title{remote_process}
\usage{
remote_process(settings)
}
\arguments{
\item{settings}{PEcAn settings list containing remotedata tags: raw_mimetype, raw_formatname, source, collection, scale, projection, qc, algorithm, credfile, pro_mimetype, pro_formatname, out_get_data, out_process_data, overwrite}
}
\description{
call rp_control (from RpTools py package) from PEcAn workflow and store the output in BETY
}
\examples{
\dontrun{
remote_process(settings)
}
}
\author{
Ayush Prasad, Istem Fer
}
