% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_process.R
\name{remotedata_db_insert}
\alias{remotedata_db_insert}
\title{Insert the output data returned from rp_control into BETYdb}
\usage{
remotedata_db_insert(
  output,
  remotefile_check_flag,
  siteid,
  out_get_data,
  out_process_data,
  write_raw_start,
  write_raw_end,
  write_pro_start,
  write_pro_end,
  raw_check,
  pro_check,
  raw_mimetype,
  raw_formatname,
  pro_mimetype,
  pro_formatname,
  dbcon
)
}
\arguments{
\item{output}{output list from rp_control}

\item{remotefile_check_flag}{remotefile_check_flag}

\item{siteid}{siteid}

\item{out_get_data}{out_get_data}

\item{out_process_data}{out_process_data}

\item{write_raw_start}{write_raw_start, start date of the raw file}

\item{write_raw_end}{write_raw_end, end date of the raw file}

\item{write_pro_start}{write_pro_start}

\item{write_pro_end}{write_pro_end}

\item{raw_check}{id, site_id, name, start_date, end_date, of the existing raw file from inputs table and file_path from dbfiles tables}

\item{pro_check}{pro_check  id, site_id, name, start_date, end_date, of the existing processed file from inputs table and file_path from dbfiles tables}

\item{raw_mimetype}{raw_mimetype}

\item{raw_formatname}{raw_formatname}

\item{pro_mimetype}{pro_mimetype}

\item{pro_formatname}{pro_formatname}

\item{dbcon}{BETYdb con}
}
\value{
list containing raw_id, raw_path, pro_id, pro_path
}
\description{
Insert the output data returned from rp_control into BETYdb
}
\examples{
\dontrun{
db_out <- remotedata_db_insert(
  output,
  remotefile_check_flag,
  siteid,
  out_get_data,
  out_process_data,
  write_raw_start,
  write_raw_end,
  write_pro_start,
  write_pro_end,
  raw_check,
  pro_check
  raw_mimetype,
  raw_formatname,
  pro_mimetype,
  pro_formatname,
  dbcon)
}
}
\author{
Ayush Prasad
}
