% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_process.R
\name{set_stage}
\alias{set_stage}
\title{set_stage}
\usage{
set_stage(result, req_start, req_end, stage)
}
\arguments{
\item{result}{dataframe containing id, site_id, name, start_date, end_date from inputs table and file_path from dbfiles table}

\item{req_start}{start date requested by the user}

\item{req_end}{end date requested by the user}

\item{stage}{the stage which needs to be set, get_remote_data or process_remote_data}
}
\value{
list containing req_start, req_end, stage, merge, write_start, write_end
}
\description{
set dates, stage and merge status for remote data download
}
\examples{
\dontrun{
raw_check <- set_stage(
  result,
  req_start,
  req_end,
  get_remote_data)
}
}
\author{
Ayush Prasad
}
