% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilC_prep.R
\name{SoilC_prep}
\alias{SoilC_prep}
\title{Prepare Soilgrids SoilC data for the SDA workflow.}
\usage{
SoilC_prep(
  Site_Info,
  Start_Date,
  End_Date,
  Time_Step = list(unit = "year", num = 1),
  outdir = NULL
)
}
\arguments{
\item{Site_Info}{Bety list of site info including site_id, lon, and lat.}

\item{Start_Date}{Start date of SDA workflow.}

\item{End_Date}{End date of SDA workflow.}

\item{Time_Step}{A list containing time step and number of time step, which allows time step to be any years or days.}

\item{outdir}{Where the final CSV file will be stored.}
}
\value{
A data frame containing AGB median and sd for each site and each time step.
}
\description{
Prepare Soilgrids SoilC data for the SDA workflow.
}
\author{
Dongchen Zhang
}
