% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soilgrids_SoilC_prep.R
\name{Soilgrids_SoilC_prep}
\alias{Soilgrids_SoilC_prep}
\title{Prepare Soilgrids SoilC data for the SDA workflow.}
\usage{
Soilgrids_SoilC_prep(
  site_info,
  start_date,
  end_date,
  timestep = list(unit = "year", num = 1),
  outdir = NULL
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{start_date}{Start date of SDA workflow.}

\item{end_date}{End date of SDA workflow.}

\item{timestep}{A list containing time step and number of time step, which allows time step to be any years or days.}

\item{outdir}{Where the final CSV file will be stored.}
}
\value{
A data frame containing AGB median and sd for each site and each time step.
}
\description{
Prepare Soilgrids SoilC data for the SDA workflow.
}
\author{
Dongchen Zhang
}
