% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_try.R
\name{try_trait_mapping}
\alias{try_trait_mapping}
\title{Create a predefined trait mapping from TRY to PEcAn}
\usage{
try_trait_mapping()
}
\value{
A named character vector where names are TRY trait names and values are PEcAn variable names.
}
\description{
Create a predefined trait mapping from TRY to PEcAn
}
