% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_MODIS.R
\name{call_MODIS}
\alias{call_MODIS}
\title{call_MODIS}
\usage{
call_MODIS(outfolder = ".", fname = "m_data.nc", start, end, lat, lon,
  size = 0, product = "MOD15A2", band = "Lai_1km", qc_band = NA,
  sd_band = NA)
}
\arguments{
\item{start}{first date in year and day-of-year. For example May 1 2010 would be 2010121}

\item{end}{laste date in year and day-of-year. For example May 1 2010 would be 2010121}

\item{lat}{Latitude of the pixel}

\item{lon}{Longitude of the pixel

depends on a number of Python libraries. sudo -H pip install numpy suds netCDF4}
}
\description{
Get MODIS data by date and location
}
\examples{
\\donotrun{
test <- call_MODIS(start="2001001",end="2016366",lat=44.0646,lon=-71.28808,size=3,qc_band = "FparLai_QC",sd_band = "LaiStdDev_1km")
}

}
