% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_MODIS.R
\name{call_MODIS}
\alias{call_MODIS}
\title{call_MODIS}
\usage{
call_MODIS(outfolder = ".", start_date, end_date, lat, lon, size = 0,
  product, band, band_qc = "", band_sd = "",
  package_method = "MODISTools")
}
\arguments{
\item{outfolder}{where the output file will be stored}

\item{start_date}{string value for beginning of date range for download in unambiguous date format (YYYYJJJ)}

\item{end_date}{string value for end of date range for download in unambiguous date format (YYYYJJJ)}

\item{lat}{Latitude of the pixel}

\item{lon}{Longitude of the pixel}

\item{size}{kmAboveBelow and kmLeftRight distance in km to be included}

\item{product}{string value for MODIS product number}

\item{band}{string value for which measurement to extract}

\item{band_qc}{string value for which quality control band, or use "NA" if you do not know or do not need QC information (optional)}

\item{band_sd}{string value for which standard deviation band, or use "NA" if you do not know or do not need StdDev information (optional)}

\item{package_method}{string value to inform function of which package method to use to download modis data. Either "MODISTools" or "reticulate" (optional)

depends on a number of Python libraries. sudo -H pip install numpy suds netCDF4 json
depends on the MODISTools package version 1.1.0}
}
\description{
Get MODIS data by date and location
}
\examples{
\dontrun{
test_modistools <- call_MODIS(product = "MOD15A2H", band = "Lai_500m", start_date = "2004300", end_date = "2004365", lat = 38, lon = -123, size = 0, band_qc = "FparLai_QC", band_sd = "LaiStdDev_500m", package_method = "MODISTools")
plot(lubridate::yday(test_modistools$calendar_date), test_modistools$data, type = 'l', xlab = "day of year", ylab = test_modistools$band[1])
test_reticulate <- call_MODIS(product = "MOD15A2H", band = "Lai_500m", start_date = "2004300", end_date = "2004365", lat = 38, lon = -123, size = 0, band_qc = "",band_sd = "", package_method = "reticulate")
}

}
\author{
Bailey Morrison
}
