% Generated by roxygen2 (4.0.1): do not edit by hand
\name{call_MODIS}
\alias{call_MODIS}
\title{call_MODIS}
\usage{
call_MODIS(outfolder = ".", fname = "m_data.nc", start, end, lat, lon,
  size = 0, product = "MOD15A2", band = "Lai_1km", qc_band = NA,
  sd_band = NA)
}
\arguments{
\item{start}{first date in year and day-of-year. For example May 1 2010 would be 2010121}

\item{end}{laste date in year and day-of-year. For example May 1 2010 would be 2010121}

\item{lat}{Latitude of the pixel}

\item{lon}{Longitude of the pixel}
}
\description{
Get MODIS data by date and location
}

