% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASA_DAAC_download.R
\name{DAAC_Set_Credential}
\alias{DAAC_Set_Credential}
\title{Set NASA DAAC credentials to the current environment.}
\usage{
DAAC_Set_Credential(replace = FALSE, folder.path = NULL)
}
\arguments{
\item{replace}{Boolean: determine if we want to replace the current credentials from the environment. The default is FALSE.}

\item{folder.path}{Character: physical path to the folder that contains the credential file. The default is NULL.}
}
\description{
Set NASA DAAC credentials to the current environment.
}
\author{
Dongchen Zhang
}
