% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.thredds.R
\name{download.thredds.AGB}
\alias{download.thredds.AGB}
\title{download.thredds.AGB}
\usage{
download.thredds.AGB(
  outdir = NULL,
  site_ids,
  run_parallel = FALSE,
  ncores = NULL
)
}
\arguments{
\item{outdir}{Where to place output}

\item{site_ids}{What locations to download data at?}

\item{run_parallel}{Logical. Download and extract files in parallel?}

\item{ncores}{Optional. If run_parallel=TRUE how many cores to use?  If left as NULL will select max number -1}
}
\value{
data.frame summarize the results of the function call
}
\description{
download.thredds.AGB
}
\examples{
\dontrun{
outdir <- "~/scratch/abg_data/"
results <- PEcAn.data.remote::download.thredds.AGB(outdir=outdir, 
           site_ids = c(676, 678, 679, 755, 767, 1000000030, 1000000145, 1000025731), 
           run_parallel = TRUE, ncores = 8)
}
}
\author{
Bailey Morrison
}
