% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandTrendr.AGB.R
\name{extract.LandTrendr.AGB}
\alias{extract.LandTrendr.AGB}
\title{extract.LandTrendr.AGB}
\usage{
extract.LandTrendr.AGB(coords, buffer = NULL, fun = "mean",
  data_dir = NULL, con = NULL, output_file = NULL, ...)
}
\arguments{
\item{coords}{vector of BETYdb site IDs or a data frame containing elements 'lon' and 'lat'}

\item{buffer}{Optional. operate over desired buffer area}

\item{fun}{Optional function to apply to buffer area.  Default - mean}

\item{data_dir}{directory where input data is located. Can be NUL if con is specified}

\item{con}{connection to PEcAn database. Can be NULL if data_dir is specified}

\item{output_file}{Path to save LandTrendr_AGB_output.RData file containing the output extraction list (see return)}
}
\value{
list of two containing the median AGB values per pixel and the corresponding standard deviation values (uncertainties)

##' @examples
\dontrun{
bety <- list(user='bety', password='bety', host='localhost',
dbname='bety', driver='PostgreSQL',write=TRUE)
con <- PEcAn.DB::db.open(bety)
bety$con <- con

data_dir <- "~/scratch/
coords <- c(2000000023,1000025731)  # US-CZ3, US-SSH

coords <- data.frame(lon=c(-90.8,91.8), lat=c(42.5,43.5))

}
}
\description{
extract.LandTrendr.AGB
}
\author{
Shawn Serbin
}
