% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.thredds.R
\name{extract_thredds_nc}
\alias{extract_thredds_nc}
\title{extract_thredds_nc}
\usage{
extract_thredds_nc(site_info, url_info, run_parallel)
}
\arguments{
\item{site_info}{list of information with the site_id, site_info, lat, lon, and time_zone. Derived from BETY using a PEcAn .xml settings file with site information. Can use the get_site_info function to generate this list.}

\item{run_parallel}{T or F option to extra data in parallel.}

\item{url}{a THREDDS url of a .nc file to extract data from.}
}
\value{
a dataframe with the values for each date/site combination from a THREDDS file
}
\description{
extract_thredds_nc
}
\author{
Bailey Morrison
}
