% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_try.R
\name{format_try_for_ma}
\alias{format_try_for_ma}
\title{Convert TRY database data to PEcAn format for meta-analysis}
\usage{
format_try_for_ma(try_data, trait_map = try_trait_mapping(), species_map = NULL)
}
\arguments{
\item{try_data}{A data frame containing data from the TRY database.}

\item{trait_map}{A named character vector for mapping TRY TraitName to PEcAn vname. 
Names should be TRY TraitName and values should be PEcAn vname. 
Defaults to \code{try_trait_mapping()}.}

\item{species_map}{An optional named vector mapping TRY \code{SpeciesName} to PEcAn BETY \code{specie_id}.
If provided, this is used to assign \code{specie_id} instead of using the raw \code{AccSpeciesID}.
If omitted, the returning data will include a \code{species_name} column to help users map PFTs later.}
}
\value{
A data frame formatted similarly to BETYdb output to be passed to \code{PEcAn.MA::jagify}.
}
\description{
This function converts trait data from the external TRY database into the tabular 
format required by the PEcAn meta-analysis module.
}
\details{
The resulting data frame is in the format returned by the PEcAn database (\code{PEcAn.DB::query.traits()}). 
To be used in \code{pecan.ma()}, \strong{it must first be run through the \code{jagify()} function} 
to ensure all missing values and columns are appropriately processed, as the meta-analysis 
model requires data converted into specific distributions and error formats. 
\code{jagify} will drop unused TRY-specific columns and compute the final \code{obs.prec} values.
}
