% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{GEDI_AGB_download}
\alias{GEDI_AGB_download}
\title{Download GEDI AGB data for the GEDI AGB extract function.}
\usage{
GEDI_AGB_download(
  start_date,
  end_date,
  outdir,
  extent,
  nfile.min = 0,
  nrow.min = 0,
  gradient = 0
)
}
\arguments{
\item{start_date}{start date (date with YYYY-MM-DD format) for downloading GEDI AGB from remote database.}

\item{end_date}{end date (date with YYYY-MM-DD format) for downloading GEDI AGB from remote database.}

\item{outdir}{Directory where the final CSV file will be stored.}

\item{extent}{the XY box (in degrees) for downloading GEDI AGB file.}

\item{nfile.min}{the minimum required file number to be downloaded and extracted, default is 0.}

\item{nrow.min}{the minimum required observation number to be extracted, default is 0.}

\item{gradient}{the gradient for iteratively enlarge the extent if the nfile.min or nrow.min are not reached, default is 0. If nfile.min or nrow.min is 0 this will be skipped.}
}
\value{
A data frame containing AGB and sd for the target spatial and temporal extent.
}
\description{
Download GEDI AGB data for the GEDI AGB extract function.
}
\author{
Dongchen Zhang
}
