% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{GEDI_AGB_prep}
\alias{GEDI_AGB_prep}
\title{Prepare GEDI AGB data for the SDA workflow.}
\usage{
GEDI_AGB_prep(
  site_info,
  time_points,
  outdir = NULL,
  buffer = 0.01,
  search_window = "3 month"
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{time_points}{A vector contains each time point within the start and end date.}

\item{outdir}{Where the final CSV file will be stored.}

\item{buffer}{buffer distance for locate GEDI AGB searching box (default is 0.01 ~ 1 km)}

\item{search_window}{search window for locate available GEDI AGB values.}
}
\value{
A data frame containing AGB and sd for each site and each time step.
}
\description{
Prepare GEDI AGB data for the SDA workflow.
}
\examples{
\dontrun{
settings_dir <- "/projectnb/dietzelab/dongchen/anchorSites/SDA/pecan.xml"
settings <- PEcAn.settings::read.settings(settings_dir)
site_info <- settings \%>\% 
  purrr::map(~.x[['run']] ) \%>\% 
  purrr::map('site')\%>\% 
  purrr::map(function(site.list){
    #conversion from string to number
    site.list$lat <- as.numeric(site.list$lat)
    site.list$lon <- as.numeric(site.list$lon)
    list(site_id=site.list$id, lat=site.list$lat, lon=site.list$lon, site_name=site.list$name)
  })\%>\% 
  dplyr::bind_rows() \%>\% 
  as.list()
  time_points <- seq(as.Date("2019-07-15"), as.Date("2023-07-15"), by = "1 year")
  outdir <- "/projectnb/dietzelab/dongchen/anchorSites/GEDI_Tami/test/"
GEDI_AGB <- GEDI_AGB_prep(site_info, time_points, outdir, buffer)
}
}
\author{
Dongchen Zhang
}
