% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.thredds.AVHRR.R,
%   R/download.thredds.AVHRR.monthAGG.R, R/download.thredds.R
\name{get_site_info}
\alias{get_site_info}
\title{get_site_info}
\usage{
get_site_info(xmlfile)

get_site_info(xmlfile)

get_site_info(xmlfile)
}
\arguments{
\item{xmlfile}{full path to pecan xml settings file}

\item{xmlfile}{full path to pecan xml settings file}

\item{xmlfile}{full path to pecan xml settings file}
}
\value{
a list of site information derived from BETY using a pecan .xml settings file with site_id, site_name, lat, lon, and time_zone.

a list of site information derived from BETY using a pecan .xml settings file with site_id, site_name, lat, lon, and time_zone.

a list of site information derived from BETY using a pecan .xml settings file with site_id, site_name, lat, lon, and time_zone.
}
\description{
get_site_info

get_site_info

get_site_info
}
\examples{
\dontrun{
xmlfile <- the full path to a pecan .xml settings file.

site_info <- get_site_info(xmlfile = "/data/bmorrison/sda/lai/pecan_MultiSite_SDA_LAI_AGB_8_Sites_2009.xml")
 }          
\dontrun{
xmlfile <- the full path to a pecan .xml settings file.

site_info <- get_site_info(xmlfile = "/data/bmorrison/sda/lai/pecan_MultiSite_SDA_LAI_AGB_8_Sites_2009.xml")
 }          
\dontrun{
xmlfile <- the full path to a pecan .xml settings file.

results <- PEcAn.data.remote::download.thredds.AGB(outdir=outdir, 
           site_ids = c(676, 678, 679, 755, 767, 1000000030, 1000000145, 1000025731), 
           run_parallel = TRUE, ncores = 8)
}
}
\author{
Bailey Morrison

Bailey Morrison

Bailey Morrison
}
