% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASA_DAAC_download.R
\name{getnetrc}
\alias{getnetrc}
\title{Set NASA DAAC credentials to the .netrc file.}
\usage{
getnetrc(dl_dir)
}
\arguments{
\item{dl_dir}{Character: physical path to the folder that the .netrc file will be generated.}
}
\description{
Set NASA DAAC credentials to the .netrc file.
}
\author{
Dongchen Zhang
}
