% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDI_AGB_prep.R
\name{getnetrc}
\alias{getnetrc}
\title{Function for building netrc file with access credentials}
\usage{
getnetrc(dl_dir)
}
\arguments{
\item{dl_dir}{Directory where the netrc file will be stored.}
}
\value{
file path of the netrc file.
}
\description{
Function for building netrc file with access credentials
}
