% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS_LAI_prep.R
\name{MODIS_LAI_prep}
\alias{MODIS_LAI_prep}
\title{Prepare MODIS LAI data for the SDA workflow.}
\usage{
MODIS_LAI_prep(
  site_info,
  time_points,
  outdir = NULL,
  search_window = 30,
  export_csv = FALSE
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{time_points}{A vector contains each time point within the start and end date.}

\item{outdir}{Where the final CSV file will be stored.}

\item{search_window}{search window for locate available LAI values.}

\item{export_csv}{Decide if we want to export the CSV file.}
}
\value{
A data frame containing LAI and sd for each site and each time step.
}
\description{
Prepare MODIS LAI data for the SDA workflow.
}
\author{
Dongchen Zhang
}
