% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS_LC_prep.R
\name{MODIS_LC_prep}
\alias{MODIS_LC_prep}
\title{Prepare MODIS land cover data for the SDA workflow.}
\usage{
MODIS_LC_prep(site_info, time_points, outdir = NULL, qc.filter = FALSE)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{time_points}{A vector contains each time point within the start and end date.}

\item{outdir}{Where the final CSV file will be stored.}

\item{qc.filter}{decide if we want to filter data by the qc band.}
}
\value{
A data frame containing MODIS land cover types for each site and each time step.
}
\description{
Prepare MODIS land cover data for the SDA workflow.
}
\author{
Dongchen Zhang
}
