% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASA_DAAC_download.R
\name{NASA_DAAC_URL}
\alias{NASA_DAAC_URL}
\title{Create URL that can be used to request data from NASA DAAC server.}
\usage{
NASA_DAAC_URL(
  base_url = "https://cmr.earthdata.nasa.gov/search/granules.json?pretty=true",
  provider,
  page_size = 2000,
  page = 1,
  concept_id,
  bbox,
  daterange = NULL
)
}
\arguments{
\item{base_url}{Character: base URL for the CMR search. 
default is "https://cmr.earthdata.nasa.gov/search/granules.json?pretty=true".}

\item{provider}{Character: ID of data provider from NASA DAAC. See `NASA_CMR_finder` for more details.}

\item{page_size}{Numeric: maximum requested length, default is 2000.}

\item{page}{Numeric: which page of the URL, default is 1.}

\item{concept_id}{Character: CMR Concept ID. See `NASA_CMR_finder` for more details.}

\item{bbox}{Numeric: vector of bounding box coordinates.}

\item{daterange}{Character: vectors of the requested start and end dates. In the form "yyyy-mm-dd".}
}
\value{
A character of URL that can be used to request data.
}
\description{
Create URL that can be used to request data from NASA DAAC server.
}
\examples{
\dontrun{
provider <- "ORNL_CLOUD"
concept_id <- "C2770099044-ORNL_CLOUD"
bbox <- "-121,33,-117,35"
daterange <- c("2022-02-23", "2022-05-30")
URL <- NASA_DAAC_URL(provider = provider, 
concept_id = concept_id, 
bbox = bbox, 
daterange = daterange)
}
}
\author{
Dongchen Zhang
}
