% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_process.R
\name{remotedata_db_check}
\alias{remotedata_db_check}
\title{remotedata_db_check}
\usage{
remotedata_db_check(
  raw_file_name,
  pro_file_name,
  start,
  end,
  siteid,
  siteid_short,
  out_get_data,
  algorithm,
  out_process_data,
  overwrite,
  dbcon
)
}
\arguments{
\item{raw_file_name}{raw_file_name}

\item{pro_file_name}{pro_file_name}

\item{start}{start date requested by user}

\item{end}{end date requested by the user}

\item{siteid}{siteid of the site}

\item{siteid_short}{short form of the siteid}

\item{out_get_data}{out_get_data}

\item{algorithm}{algorithm}

\item{out_process_data}{out_process_data}

\item{overwrite}{overwrite}

\item{dbcon}{BETYdb con}
}
\value{
list containing remotefile_check_flag, start, end, stage_get_data, write_raw_start, write_raw_end, raw_merge, existing_raw_file_path, stage_process_data, write_pro_start, write_pro_end, pro_merge, input_file, existing_pro_file_path, raw_check, pro_check
}
\description{
check the status of the requested data in the DB
}
\examples{
\dontrun{
dbstatus <- remotedata_db_check(
  raw_file_name,
  pro_file_name,
  start,
  end,
  siteid,
  siteid_short,
  out_get_data,
  algorithm,
  out_process_data,
  overwrite
  dbcon)
}
}
\author{
Ayush Prasad
}
