% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMAP_SMP_prep.R
\name{SMAP_SMP_prep}
\alias{SMAP_SMP_prep}
\title{Prepare SMAP Soil Moisture (SMP) data for the SDA workflow.}
\usage{
SMAP_SMP_prep(
  site_info,
  start_date,
  end_date,
  time_points,
  outdir,
  search_window = 30,
  export_csv = TRUE,
  update_csv = FALSE
)
}
\arguments{
\item{site_info}{Bety list of site info including site_id, lon, and lat.}

\item{start_date}{Start date of SDA workflow.}

\item{end_date}{End date of SDA workflow.}

\item{time_points}{A vector contains each time point within the start and end date.}

\item{outdir}{Where the final CSV file, and the CSV file from GEE are stored.}

\item{search_window}{search window for locate available SMP values.}

\item{export_csv}{Decide if we want to export the CSV file.}

\item{update_csv}{Decide if we want to update current CSV file given an updated SMAP_gee.csv file}
}
\value{
A data frame containing SMAP smp and sd for each site and each time step.
}
\description{
Prepare SMAP Soil Moisture (SMP) data for the SDA workflow.
}
\author{
Dongchen Zhang
}
